Created the first working version of a program designed to generate examples for the periodic graphs
The current version is a simple implementation of a program that will check all the subgraphs of a dense periodic graph with a given number of vertices in the fundamental domain and a given number of actions. It will record the maximal subgraphs whose corresponding newton polytopes have the same volume as the dense version. This program needs to be changed to write results to a file rather than display them to the user. Additionally, the program still needs to be checked for correctness and can likely be optimized in many areas.