Attempted to multi thread the subgraphs program with limited success. Previous version saved for backup
Subgraphs.m2 now contains a multi threaded version of the subgraphs program. Have had limited success with this program. It will occasionally halt execution claiming a "bus error". However, when it does run, it is significantly faster (takes about 60% of the time) than the regular subgraphs program and correctly records the results. Need to spend some time attempting to fix the occasional crashes.