Fixed some prevalent bugs in streamlinedFaces.m2 and recreated data
Fixed 2 important issues in streamlinedFaces.m2. First, initially, the program would check the dimensions of the generated ideals until it reached the first that was not -1. Then, it would output the corresponding graph to one of the two output files depending on whether or not it was a false positive for solutions on that facet. Now, the program should check all dimensions before confirming that there are no solutions on the facets. The second bug concerned the fact that the z variable was allowed to be 0. This could produce false positives on some of the facets. In order to prevent this, the variable zi is introduced to ensure that z must be invertible.