Added some additional conditions to check for in the program
Now checks that no two permutations on a given graph fix the same vertex (if they do there are solutions on a facet and there is no need to check that graph). Additionally, checks that no three permutations (a permutation may be exchanged with its inverse for this check) send a vertex to the same location. As with the previous case, this would mean there are solutions on a facet and there is no need to check this graph.